Release v3.0.0
Rewrite Snipping Lens. Snipping Lens 3 is a more polished and responsive thanks to libraries such as watchdog, PySide6, flet, and copykitten.

Snipping Lens 3 runs Python directly without the need to contain it inside an .exe. This solves an issue where antiviruses would detect the app as harmful (false positive).

This release is built only for Windows for now as the majority of the users are on WIndows. Linux release are coming soon.

Users do not need Python installed to run the application as Snipping Lens 3 uses its own Python environment that does not interfere with the system. Everything is contained inside the project.

For Windows, simply run setup_win.vbs and use the generated shortcut to run the app. The app is entirely portable, simply move the project to your desired folder and re-run setup_win.vbs to generate the shortcut for the app.